Add Base Crypto Invoicer
