+++ lib/payment_services/base.rb
@@ -6,7 +6,7 @@ module PaymentServices
   # Базовый класс для платежного сервиса. Описывает подсервисы и хранит конфигурацию
   #
   class Base
-    SUBSERVICES = %i[invoicer importer payout_adapter client].freeze
+    SUBSERVICES = %i[invoicer importer payout_adapter client crypto_invoice].freeze
 
     # Реестр подсервисов
     class Registry
